api/pyproject.toml
[tool.poetry]
name = "artigo_web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.3"
django-filter = "^2.4.0"
Markdown = "^3.3.4"
psycopg2-binary = "^2.8.6"
djangorestframework-jwt = "^1.11.0"
python-dateutil = "^2.8.1"
celery = "^5.0.5"
redisai = "^1.0.1"
pandas = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.5.2"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
